pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arc-stack"
version = "0.1.3"
description = "Stack is a single-cell foundation model that enables in-context learning at inference time."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Mingze Dong", email = "mingze.dong@arcinstitute.org" },
    { name = "Yusuf Roohani", email = "yusuf.roohani@arcinstitute.org" },
]
requires-python = ">=3.9"
dependencies = [
    "anndata>=0.9",
    "h5py>=3.8",
    "geomloss>=0.2.6",
    "numpy>=1.22",
    "pandas>=1.5",
    "psutil>=5.9",
    "pytorch-lightning>=2.1",
    "PyYAML>=6.0",
    "scipy>=1.8",
    "scvi-tools>=1.0",
    "torch>=2.0",
    "wandb>=0.15",
]

[project.urls]
Homepage = "https://github.com/ArcInstitute/stack"
Source = "https://github.com/ArcInstitute/stack"

[project.scripts]
stack-train = "stack.cli.launch_training:main"
stack-finetune = "stack.cli.launch_finetuning:main"
stack-embedding = "stack.cli.embedding:main"
stack-generation = "stack.cli.generation:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["stack*"]
